+++ examples/client/validate_srv_record/validate_srv_record.go
@@ -9,7 +9,6 @@ import (
 )
 
 func main() {
-
 	// Load the client
 	client, err := paymail.NewClient()
 	if err != nil {
@@ -20,17 +19,17 @@ func main() {
 	var srv *net.SRV
 	srv, err = client.GetSRVRecord(paymail.DefaultServiceName, paymail.DefaultProtocol, "moneybutton.com")
 	if err != nil {
-		log.Fatal("error getting SRV record: " + err.Error())
+		log.Fatalf("error getting SRV record: %s", err.Error())
 	}
 
 	// Found record!
-	log.Println("found SRV record:", srv)
+	log.Printf("found SRV record: %v", srv)
 
 	// Validate the record (1 instead of 10, moneybutton deviated from the defaults)
 	err = client.ValidateSRVRecord(
 		context.Background(), srv, paymail.DefaultPort, 1, paymail.DefaultWeight,
 	)
 	if err != nil {
-		log.Fatal("failed validating SRV record: " + err.Error())
+		log.Fatalf("failed validating SRV record: " + err.Error())
 	}
 }