+++ examples/server/run_server/run_server.go
@@ -1,35 +1,70 @@
 package main
 
 import (
-	"log"
+	"fmt"
+	"net/http"
 	"time"
+
+	"github.com/gin-gonic/gin"
+
+	"github.com/bitcoin-sv/go-paymail/logging"
 
 	"github.com/bitcoin-sv/go-paymail/server"
 )
 
 func main() {
+	logger := logging.GetDefaultLogger()
 
 	// initialize the demo database
 	if err := InitDemoDatabase(); err != nil {
-		log.Fatal(err.Error())
+		logger.Fatal().Msg(err.Error())
 	}
+
+	sl := server.PaymailServiceLocator{}
+	sl.RegisterPaymailService(new(demoServiceProvider))
+	sl.RegisterPikeService(new(demoServiceProvider))
 
 	// Custom server with lots of customizable goodies
 	config, err := server.NewConfig(
-		new(demoServiceProvider),
+		&sl,
 		server.WithBasicRoutes(),
-		server.WithDomain("localhost"), // todo: make this work locally?
+		server.WithDomain("localhost"),
 		server.WithDomain("another.com"),
 		server.WithDomain("test.com"),
 		server.WithGenericCapabilities(),
 		server.WithPort(3000),
 		server.WithServiceName("BsvAliasCustom"),
 		server.WithTimeout(15*time.Second),
+		server.WithCapabilities(customCapabilities()),
 	)
+	config.Prefix = "http://" //normally paymail requires https, but for demo purposes we'll use http
 	if err != nil {
-		log.Fatal(err.Error())
+		logger.Fatal().Msg(err.Error())
 	}
 
 	// Create & start the server
-	server.StartServer(server.CreateServer(config))
+	server.StartServer(server.CreateServer(config), config.Logger)
+}
+
+func customCapabilities() map[string]any {
+	exampleBrfcKey := "406cef0ae2d6"
+	return map[string]any{
+		"custom_static_boolean": false,
+		"custom_static_int":     10,
+		exampleBrfcKey:          true,
+		"custom_callable_cap": server.CallableCapability{
+			Path:   fmt.Sprintf("/display_paymail/%s", server.PaymailAddressTemplate),
+			Method: http.MethodGet,
+			Handler: func(c *gin.Context) {
+				incomingPaymail := c.Param(server.PaymailAddressParamName)
+
+				response := map[string]string{
+					"paymail": incomingPaymail,
+				}
+
+				c.Header("Content-Type", "application/json")
+				c.JSON(http.StatusOK, response)
+			},
+		},
+	}
 }