+++ server/interface.go
@@ -4,7 +4,37 @@ import (
 	"context"
 
 	"github.com/bitcoin-sv/go-paymail"
+	"github.com/bitcoin-sv/go-paymail/spv"
 )
+
+type PaymailServiceLocator struct {
+	paymailService PaymailServiceProvider
+	pikeService    PikeServiceProvider
+}
+
+func (l *PaymailServiceLocator) RegisterPaymailService(s PaymailServiceProvider) {
+	l.paymailService = s
+}
+
+func (l *PaymailServiceLocator) GetPaymailService() PaymailServiceProvider {
+	if l.paymailService == nil {
+		panic("PaymailServiceProvider was not registered")
+	}
+
+	return l.paymailService
+}
+
+func (l *PaymailServiceLocator) RegisterPikeService(s PikeServiceProvider) {
+	l.pikeService = s
+}
+
+func (l *PaymailServiceLocator) GetPikeService() PikeServiceProvider {
+	if l.pikeService == nil {
+		panic("PikeServiceProvider was not registered")
+	}
+
+	return l.pikeService
+}
 
 // PaymailServiceProvider the paymail server interface that needs to be implemented
 type PaymailServiceProvider interface {
@@ -36,6 +66,14 @@ type PaymailServiceProvider interface {
 
 	VerifyMerkleRoots(
 		ctx context.Context,
-		merkleProofs []string,
+		merkleProofs []*spv.MerkleRootConfirmationRequestItem,
+	) error
+}
+
+type PikeServiceProvider interface {
+	AddContact(
+		ctx context.Context,
+		requesterPaymail string,
+		contact *paymail.PikeContactRequestPayload,
 	) error
 }